rag-evaluation/requirements.txt
ragas
datasets
requests
python-dotenv
pandas
langchain-openai
langchain-community
sentence-transformers
